Just changed some script names
